Be less dumb
There was an odd occurance I've probably spent 3 hours tracking down.
Reserved layouts seemed affected by the position of other layouts. That
wouldn't be unusual if everything was normally positioned, but this
layout is absolute.

Turns out it was not absolute. Literally. That was the only problem. I
forgot to put `position: absolute` on a container.

My face is still sore from that facepalm.