Better APK url parsing
The build:status command seems to output different stuff every once in a
while. The last build triggered an incorrect slack message linking to
the word "APK" instead of the actual url, since it was printing the
wrong column.

This actually looks for a url-like thing and only prints that. Hopefully
it's more robust to changes.