Refetch groups periodically in dashboard mode
Run a query occasionally if in dashboard mode. Fetches all groups and
their on state. This is pretty expensive though, so it's a short term
fix until I get gql subscriptions configured. Also, turns out the
network is horribly unreliable on mobile. Maybe I've misconfigured
something, but that'd be a weird bug.